+++ include/client.h
@@ -224,6 +224,77 @@ class Client : public SRObject
                         const SRMemoryLayout mem_layout);
 
         /*!
+        *   \brief Puts a byte string into the database
+        *   \details The provided byte string is placed into the database
+        *            without any type associated with the bytes.  
+        *            The final key under which the bytes are stored
+        *            may be formed by applying a prefix to the supplied
+        *            name. See use_bytes_ensemble_prefix() for
+        *            more details.
+        *   \param name The name for referencing the bytes in the database
+        *   \param data The data
+        *   \param n_bytes The number of bytes in the provided data
+        *   \throw SmartRedis::Exception if put bytes command fails
+        */
+        void put_bytes(const std::string& name,
+                       const void* bytes,
+                       const size_t n_bytes);
+
+        /*!
+        *   \brief Retrieve bytes from the database and place into a new
+        *          memory buffer
+        *   \details The key used to locate the stored bytes
+        *            may be formed by applying a prefix to the supplied
+        *            name. See set_data_source()
+        *            and use_bytes_ensemble_prefix() for more details.
+        *            Unlike other "get" operations in the client, the
+        *            the byte data memeory is allocated but not managed
+        *            by the client.  The caller is responsible for invoking
+        *            free() on the returned memory.
+        *   \param name The name for referencing the bytes in the database
+        *   \param data A user-provided pointer that will be modified to
+        *               point to a new block of memory containing the bytes
+        *   \param n_bytes A user-provided size_t variable that will be modified
+        *                  to indicate the total number of bytes.
+        *   \throw SmartRedis::Exception if get bytes command fails
+        */
+        void get_bytes(const std::string& name,
+                       void*& data,
+                       size_t& n_bytes);
+
+        /*!
+        *   \brief Retrieve bytes from the database and place into memory
+        *          provided by the caller
+        *   \details The key used to locate the stored bytes
+        *            may be formed by applying a prefix to the supplied
+        *            name. See set_data_source()
+        *            and use_bytes_ensemble_prefix() for more details.
+        *   \param name The name for referencing the bytes in the database
+        *   \param data A buffer into which to place byte data
+        *   \param n_bytes The number of bytes in the provided memory block
+        *   \param n_used_bytes The number of bytes in the provided 
+        *                       buffer that were used
+        *   \throw SmartRedis::Exception if unpack bytes command fails or
+        *          the number of provided bytes is less than the retrieved
+        *          number of bytes
+        */
+        void unpack_bytes(const std::string& name,
+                          void* data,
+                          const size_t n_bytes,
+                          size_t& n_used_bytes);
+
+        /*!
+        *   \brief Delete bytes from the database
+        *   \details The key used to locate the bytes to be
+        *            deleted may be formed by applying a prefix to the
+        *            supplied name. See set_data_source()
+        *            and use_bytes_ensemble_prefix() for more details.
+        *   \param name The name of the bytes to delete
+        *   \throw SmartRedis::Exception if delete bytes command fails
+        */
+        void delete_bytes(const std::string& name);
+
+        /*!
         *   \brief Retrieve the tensor data, dimensions, and type for the
         *          provided tensor key. This function will allocate and retain
         *          management of the memory for the tensor data.
@@ -796,18 +867,6 @@ class Client : public SRObject
         bool key_exists(const std::string& key);
 
         /*!
-        *   \brief Check if a model (or script) key exists in the database
-        *   \details The model or script key used to check for existence
-        *            may be formed by applying a prefix to the supplied
-        *            name. See set_data_source()
-        *            and use_model_ensemble_prefix() for more details.
-        *   \param name The model/script name to be checked in the database
-        *   \returns Returns true if the model exists in the database
-        *   \throw SmartRedis::Exception if model exists command fails
-        */
-        bool model_exists(const std::string& name);
-
-        /*!
         *   \brief Check if a tensor key exists in the database
         *   \details The tensor key used to check for existence
         *            may be formed by applying a prefix to the supplied
@@ -830,6 +889,30 @@ class Client : public SRObject
         *   \throw SmartRedis::Exception if dataset exists command fails
         */
         bool dataset_exists(const std::string& name);
+
+        /*!
+        *   \brief Check if a model (or script) key exists in the database
+        *   \details The model or script key used to check for existence
+        *            may be formed by applying a prefix to the supplied
+        *            name. See set_data_source()
+        *            and use_model_ensemble_prefix() for more details.
+        *   \param name The model/script name to be checked in the database
+        *   \returns Returns true if the model exists in the database
+        *   \throw SmartRedis::Exception if model exists command fails
+        */
+        bool model_exists(const std::string& name);
+
+        /*!
+        *   \brief Check if bytes exists in the database
+        *   \details The key used to check for existence
+        *            may be formed by applying a prefix to the supplied
+        *            name. See set_data_source()
+        *            and use_bytes_ensemble_prefix() for more details.
+        *   \param name The bytes name to be checked in the database
+        *   \returns Returns true if the bytes exists in the database
+        *   \throw SmartRedis::Exception if bytes exists command fails
+        */
+        bool bytes_exists(const std::string& name);
 
         /*!
         *   \brief Check if a key exists in the database, repeating the check
@@ -900,6 +983,25 @@ class Client : public SRObject
         *   \throw SmartRedis::Exception if poll model command fails
         */
         bool poll_model(const std::string& name,
+                        int poll_frequency_ms,
+                        int num_tries);
+
+        /*!
+        *   \brief Check if bytes exists in the database, repeating
+        *          the check at a specified polling interval
+        *   \details The key used to check for existence
+        *            may be formed by applying a prefix to the supplied
+        *            name. See set_data_source()
+        *            and use_bytes_ensemble_prefix() for more details.
+        *   \param name The bytes name to be checked in the database
+        *   \param poll_frequency_ms The time delay between checks,
+        *                            in milliseconds
+        *   \param num_tries The total number of times to check for the name
+        *   \returns Returns true if the bytes is found within the
+        *            specified number of tries, otherwise false.
+        *   \throw SmartRedis::Exception if poll bytes command fails
+        */
+        bool poll_bytes(const std::string& name,
                         int poll_frequency_ms,
                         int num_tries);
 
@@ -1010,6 +1112,25 @@ class Client : public SRObject
         *         aggregation list prefixing
         */
         void use_list_ensemble_prefix(bool use_prefix);
+
+        /*!
+        *   \brief Control whether bytes are prefixed
+        *   \details This function can be used to avoid key collisions in an
+        *            ensemble by prepending the string value from the
+        *            environment variable SSKEYIN and/or SSKEYOUT to
+        *            raw byte names.  Prefixes will only be used if
+        *            they were previously set through the environment variables
+        *            SSKEYOUT and SSKEYIN. Keys for raw bytes created
+        *            before this function is called will not be retroactively
+        *            prefixed. By default, the client prefixes raw bytes
+        *            keys with the first prefix specified with the SSKEYIN
+        *            and SSKEYOUT environment variables.
+        *  \param use_prefix If set to true, all future operations
+        *                    on raw bytes will use a prefix, if available.
+        *  \throw SmartRedis::Exception for failed activation of
+        *         raw byte prefixing
+        */
+        void use_bytes_ensemble_prefix(bool use_prefix);
 
         /*!
         *   \brief Returns information about the given database node
@@ -1570,6 +1691,12 @@ class Client : public SRObject
         bool _use_list_prefix;
 
         /*!
+        * \brief Flag determining whether prefixes should be used
+        *        for raw byte keys.
+        */
+        bool _use_bytes_prefix;
+
+        /*!
         * \brief Our configuration options, used to access runtime settings
         */
         ConfigOptions* _cfgopts;
@@ -1664,6 +1791,15 @@ class Client : public SRObject
         */
         inline std::string _build_list_key(const std::string& list_name,
                                            bool on_db);
+        /*!
+        * \brief Build full formatted key for bytes, based on
+        *        current prefix settings.
+        * \param name Unprefixed bytes name
+        * \param on_db Indicates whether the key refers to an entity
+        *              which is already in the database.
+        */
+        inline std::string _build_bytes_key(const std::string& name,
+                                            bool on_db);
 
         /*!
         *   \brief Append the Command associated with